llm-backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic==2.10.5
sqlalchemy==2.0.36
requests==2.32.3
